requirements.txt
Django==5.2.5
gunicorn==21.2.0
whitenoise==6.11.0
django-cloudinary-storage==0.3.0
cloudinary==1.36.0
psycopg2-binary==2.9.10
pillow==11.3.0
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.2.1